fix code review comments
